Pow opt (#104)
* WIP

* Monero changes

* WIP

* WIP

* ZCash Ethereum

* WIP

* Streamlined Stratum Receive logic

* Eliminate BigInteger parse

* Support for compact block target (bits)

* Improved dupe checks

* Don't use tasks for long-running stuff

* Move uv-socket keepAlive to client-socket where it belongs

* Disable SimultaneousAccepts

* Cleanup

* Cleanup

* Paging fallback for Monero-like daemons not supporting RPC method "transfer_split"

* Some cleanup

* Produce less garbage in API Server

* Reduced memory strain for period hashrate calculations

* Consolidate Difficulty types for database entities and schema
Fixed extra-nonce problem for Bitcoin family

* Fixed ZCash share diff

* Various fixes

* Disable console writeline

* Simplified inheritance hierarchy

* Cleanup

* Fixed tests

+++ src/MiningCore/Persistence/Postgres/Entities/Block.cs
@@ -27,7 +27,7 @@ namespace MiningCore.Persistence.Postgres.Entities
         public long Id { get; set; }
         public string PoolId { get; set; }
         public long BlockHeight { get; set; }
-        public long NetworkDifficulty { get; set; }
+        public double NetworkDifficulty { get; set; }
         public string Status { get; set; }
         public double ConfirmationProgress { get; set; }
         public double? Effort { get; set; }
